backend/requirements.txt

Django==3.1.5
django-cors-headers==3.6.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
-U djoser
Pillow==8.1.0

